requirements.txt
altgraph==0.17.4
asgiref==3.9.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
blinker==1.9.0
cbor2==5.7.1
certifi==2025.8.3
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
click==8.3.1
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
Django==5.2.6
django-cors-headers==4.9.0
django-sslserver==0.22
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Flask==3.0.0
hyperlink==21.0.0
idna==3.10
Incremental==24.11.0
itsdangerous==2.2.0
Jinja2==3.1.6
json5==0.12.1
MarkupSafe==3.0.3
msgpack==1.1.2
numpy==2.3.3
numpy-financial==1.0.0
packaging==25.0
pefile==2023.2.7
pillow==12.1.1
psutil==7.1.0
psycopg==3.2.10
psycopg2==2.9.10
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.8
PyJWT==2.10.1
pyodbc==5.2.0
pyOpenSSL==25.3.0
pysmb==1.2.13
python-barcode==0.16.1
python-dotenv==1.2.1
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.3
querystring-parser==1.2.4
redis==7.2.1
reportlab==4.4.9
requests==2.32.5
sentry-sdk==2.46.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.3
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
ujson==5.11.0
urllib3==2.5.0
user-agents==2.2.0
Werkzeug==3.0.1
whitenoise==6.10.0
zope.interface==8.1.1
